+++ src/Converter/SerializerConverterAdapter.php
@@ -8,7 +8,7 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 /**
  * @author Piotr Kugla <piku235@gmail.com>
  */
-final class SerializerObjectConverterAdapter implements ConverterInterface
+final class SerializerConverterAdapter implements ConverterInterface
 {
     private DenormalizerInterface $denormalizer;
     private array $context;
@@ -21,14 +21,6 @@ final class SerializerObjectConverterAdapter implements ConverterInterface
 
     public function convert(mixed $value, string $type): mixed
     {
-        if (!class_exists($type)) {
-            throw new \InvalidArgumentException('Conversion to class types is only supported.');
-        }
-
-        if ($value instanceof $type) {
-            return $value;
-        }
-
         try {
             return $this->denormalizer->denormalize($value, $type, null, $this->context);
         } catch (NotNormalizableValueException | \TypeError $e) {