+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Transition to Symfony v6.0
 - `ConverterManager` no longer checks whether the passed value is already of the given type. Now, each converter should
 take care of this itself.
-- `SerializerObjectConverterAdapter` won't convert a value if it's already of a given type.
+- `SerializerObjectConverterAdapter` to `SerializerConverterAdapter`. It no longer checks if the type is an object type, 
+it fully delegates to the adapted denormalizer.
 
 ### Removed
 - Annotations